backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
asttokens==3.0.0
attrs==25.3.0
cachetools==5.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.2
Django==4.2.4
django-allauth==0.58.2
django-cors-headers==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
exceptiongroup==1.2.2
executing==2.2.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.4
inflection==0.5.1
ipython==8.18.1
isodate==0.7.2
jedi==0.19.2
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
matplotlib-inline==0.1.7
oauthlib==3.2.2
openai==1.79.0
outcome==1.3.0.post0
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.31.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.3
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.25.1
rsa==4.9.1
selenium==4.32.0
sniffio==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.6.1
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
colorama==0.4.6